# specs/001-github-project-chat/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Projects Chat API
  description: |
    REST API for the GitHub Projects Chat Interface.
    Enables natural language task management for GitHub Projects V2.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /projects:
    get:
      tags: [projects]
      summary: List user's GitHub Projects
      description: Returns all projects accessible to the authenticated user
      operationId: listProjects
      security:
        - sessionCookie: []
      parameters:
        - name: refresh
          in: query
          description: Force refresh from GitHub API (bypass cache)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}:
    get:
      tags: [projects]
      summary: Get project details
      description: Returns detailed project information including items
      operationId: getProject
      security:
        - sessionCookie: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            example: PVT_kwDOABCD1234
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    Project:
      type: object
      properties:
        project_id:
          type: string
          example: PVT_kwDOABCD1234
        name:
          type: string
          example: Sprint 42
        type:
          type: string
          enum: [organization, user, repository]
        owner_login:
          type: string
          example: acme-corp
        url:
          type: string
          format: uri
        item_count:
          type: integer
        status_columns:
          type: array
          items:
            $ref: '#/components/schemas/StatusColumn'
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
    StatusColumn:
      type: object
      properties:
        field_id:
          type: string
        name:
          type: string
          example: Todo
        option_id:
          type: string
        color:
          type: string
          nullable: true
    Task:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        project_id:
          type: string
        github_item_id:
          type: string
        title:
          type: string
          maxLength: 256
        description:
          type: string
          nullable: true
          maxLength: 65535
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            detail: Session expired or invalid
            code: UNAUTHORIZED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            detail: Task not found
            code: NOT_FOUND